Fix add_snaphot method only for weighted graphs.
This commit fixes an issue with the add_snapshot method.  If you are
using the add_weighted_edges_from method it expects a 3-tuple.  Note
that this commit will likely not work if the underlying graph isn't a
weighted one, unless an unweighted graph impliclity stores a weight
value of 1 for it.